Serviço de categorias se comunica com a API e possui alguns dos principais testes

+++ FrontEnd/app/scripts/services/categories.js
@@ -18,22 +18,32 @@ angular.module('spawesomeApp')
     };
 
     dataFactory.getCategory = function (id) {
+        if(id == null)
+          throw 'Id não informado';
         return $http.get(urlBase + '/' + id);
     };
 
     dataFactory.insertCategory = function (category) {
+        if(category == null)
+          throw 'category não informada';
         return $http.post(urlBase, category);
     };
 
     dataFactory.updateCategory = function (category) {
+        if(category == null)
+          throw 'category não informada';
         return $http.put(urlBase + '/' + category.Id, category)
     };
 
     dataFactory.deleteCategory = function (id) {
+        if(id == null)
+          throw 'Id não informado';
         return $http.delete(urlBase + '/' + id);
     };
 
     dataFactory.getSubCategories = function (id) {
+        if(id == null)
+          throw 'Id não informado';
         return $http.get(urlBase + '/' + id + '/SubCategories');
     };
